+++ app/controllers/projects_controller.rb
@@ -40,6 +40,14 @@ class ProjectsController < ApplicationController
     pre_fetch_incoming_trusts(@projects)
   end
 
+  def user_in_progress
+    authorize Project
+    @pager, @projects = pagy(Project.assigned_to(current_user).in_progress.includes(:assigned_to))
+
+    pre_fetch_establishments(@projects)
+    pre_fetch_incoming_trusts(@projects)
+  end
+
   def unassigned
     authorize Project
     @pagy, @projects = pagy(policy_scope(Project.unassigned_to_user.assigned_to_regional_caseworker_team))