+++ CHANGELOG.md
@@ -19,7 +19,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   new view switches to a specific table that is focused on showing completed
   projects which is sorted by completion date
 - a new view that shows all completed projects assigned to Regional casework
-  services at `/projects/regional-casework-services/completed`
+- a new view that shows a user the project they are assigned and that are
+  in-progress at `/projects/user/in-progress`
 
 ## [Release 16][release-16]